requirements.txt
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
pymongo>=4.6.0
Pillow>=10.0.0
groq>=0.4.0
